0.2 实现sin和cos的计算

+++ src/cn/edu/ncepu/sa/activemq/Mul.java
@@ -13,8 +13,8 @@ public class Mul extends C2Component {
 		String[] tmp = msg.split(",");
 		if (tmp.length >= 3) {
 			if (tmp[0].equals("*")) {
-				int x = Integer.parseInt(tmp[1]);
-				int y = Integer.parseInt(tmp[2]);
+				Double x = Double.parseDouble(tmp[1]);
+				Double y = Double.parseDouble(tmp[2]);
 				return String.valueOf((x * y));
 			}
 		}
